Make all compilable.

+++ binary-encryption/dacryfile/src/vx_main.c
@@ -32,9 +32,9 @@
 
 #include <linux/types.h>
 #include <asm/unistd.h>
-#include <linux/fcntl.h>
+#include <fcntl.h>
 #include <linux/mman.h>
-#include <linux/dirent.h>
+#include <dirent.h>
 
 #include <dl_libv2.h>
 #include <elf.h>
@@ -165,7 +165,7 @@ vx_main (void)
 
 	ptr = (long) &entry;
 	// len = total size - (offset of _start)
-	(long)len = (long)phdr->p_filesz - ((long)ptr - (long)phdr->p_vaddr);
+	len = (long)phdr->p_filesz - ((long)ptr - (long)phdr->p_vaddr);
 
 	name[0] = '%';
 	name[1] = 'x';
